requirements.txt
streamlit
python-dotenv
rank_bm25
faiss-cpu
sentence-transformers
PyMuPDF
python-docx
beautifulsoup4
readability-lxml
requests
httpx
weasyprint
reportlab
Pillow
pytesseract
google-generativeai